Revert "Update panfork to 24.1.0-devel"
This reverts commit bb37ded8b9aec7cf79a28737a2e2ecb353365810, reversing
changes made to 299b97bb9d55e89d3c9b2f6097fb0c7c8e32192b.

+++ src/mesa/main/image.h
@@ -82,7 +82,7 @@ _mesa_image_row_stride( const struct gl_pixelstore_attrib *packing,
                         GLint width, GLenum format, GLenum type );
 
 
-extern intptr_t
+extern GLint
 _mesa_image_image_stride( const struct gl_pixelstore_attrib *packing,
                           GLint width, GLint height,
                           GLenum format, GLenum type );
